grep.[ch]: invoke "opt->output" callback via a static wrapper
Invoke the optional "opt->output" callback via a new static
grep_output() wrapper, and stop including the old "std_output" in the
GREP_OPT_INIT macro added in 72365bb4992 (grep API: call grep_config()
after grep_init(), 2022-02-16).

In a subsequent commit we'll make use of the GREP_OPT_INIT outside of
grep.c, which we couldn't do it if referred to a static function
contained within that file.

The change being made here in grep_source_1() would be introducing a
logic error if we had to deal with the case of "opt->output" being set
explicitly to the old equivalent to "std_output", but that's not the
case, it's either a custom callback or NULL.

We might want to further generalize the special-case based on whether
we're using the default fwrite(), see 08303c3636e (grep: fix coloring
of hunk marks between files, 2011-06-05) for that behavior, but doing
so would be an unrelated change.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>