revisions API: add and use a limited REV_INFO_INIT
Add a new REV_INFO_INIT macro, as noted in the added API documentation
this isn't sufficient to initialize a "struct rev_info" yet, but
subsequent commits will get us closer to that eventual goal.

Even though we can't fully initialize with this REV_INFO_INIT it's
useful to clearly separate those things that we can statically
initialize, and those that we can't.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>