git-sh-setup: don't translate require_work_tree_exists() output
Partially revert d323c6b6410 (i18n: git-sh-setup.sh: mark strings for
translation, 2016-06-17). Since d03ebd411c6 (rebase: remove the
rebase.useBuiltin setting, 2019-03-18) we've had no in-tree user of
require_work_tree_exists().

Any external user of the git-sh-setup API has had good reason to
expect that the interface messages emitted would be in the LC_ALL=C
locale, as the only passing reference to the human readable output
added in d577cd216e0 (git-sh-setup: make require_clean_work_tree part
of the interface, 2011-12-20) asked the API user to provide a
human-readable string to be interpolated as a parameter into the
middle of a sentence.

As of a preceding commit commit which added an "INTERNATIONALIZATION"
section this expectation is now explicit, so let's stop marking this
for translation.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>