CI: make "make -n artifacts-tar" easier to debug
Don't ignore the exit code from "make" by piping it directly to
"grep", let's instead use an intermediate file in the ".build/"
directory. Furthermore let's "cat" the file so we can see what went
wrong if this step fails.

A subsequent change will change how the "artifacts-tar" target is
treated when *not* run with this "make -n". This change isn't required
for that subsequent change, but only to make it easier to debug things
when they fail.

I had an earlier iteration of the subsequent patch where we'd fail in
this "make -n", and due to how we pipe STDERR to stdout and grep only
^tar out of it the CI output wouldn't help to narrow it down.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>