git-sh-i18n: remove git-sh-i18n--envsubst
Remove the sh-i18n--envsubst.c program added in my
ba67aaf2d05 (git-sh-i18n--envsubst: our own envsubst(1) for
eval_gettext(), 2011-05-14), and replace our "eval_gettext" with an
"eval_gettext_unsafe", which will "eval" the message to be emitted
instead.

The reason envsubst(1) (which git-sh-i18n--envsubst was forked from)
exists in GNU libintl is so that "eval_gettext" can be used on input
that's potentially hostile, i.e. to avoid passing it to "eval", and
because the "eval" function will expect input whose double-quotes are
escaped. We therefore couldn't have used it for any arbitrary input.

Since ba67aaf2d05 was written we've mostly moved away from using
in-tree shellscripts to implement git in favor of built-ins, and since
the preceding commit the potentially external "$LONG_USAGE" is no
longer used with "eval_gettext".

So we can just move away from "eval_gettext", and by doing so get rid
of a significant amount of shims needed to support the more general
use-case.

Let's remove both the documentation of "gettext" and "eval_gettext"
that I added in e00cf070a45 (git-sh-i18n.sh: add no-op gettext() and
eval_gettext() wrappers, 2011-05-14).

Since the subsequent 5e9637c6297 (i18n: add infrastructure for
translating Git with gettext, 2011-11-18) this documentation has been
inaccurate, which says something about the likelihood that anyone has
tried to use this out-of-tree.

As the documentation states this library is for git's own internal
use, so let's not include this level of detail in our public
documentation. We have po/README.md to describe how to use the API,
let's update that to describe these new functions, including an
example about "git-am.sh" that's been incorrect since
49eb8d39c78 (Remove contrib/examples/*, 2018-03-25).

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>