git-sh-setup: stop translating LONG_USAGE template message
When we run built-ins the "usage: " part of the message is not
translatable, see usage_builtin() in usage.c, we should ideally fix
that, but until then there's no reason to try even harder in the
shellscripts than in the C code by offering translations.

The reason not to do so is that if we don't need to deal with
arbitrary user input in "eval_gettext" we can make that function much
simpler, as will be shown in a subsequent commit. The "LONG_USAGE"
variable will potentially be provided by out-of-tree users. See the
documentation added in 850844e28f7 (Documentation/git-sh-setup.txt:
programmer's docs, 2007-01-17).

This partially reverts d323c6b6410 (i18n: git-sh-setup.sh: mark
strings for translation, 2016-06-17). Between 850844e28f7 and
d323c6b6410 (2007-2016) users of "git-sh-setup" would consistently get
LC_ALL=C messaging in this case, but after that might have
unexpectedly started getting translated messages. Let's document our
non-promises in this regard.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>