quote API: remove strbuf_detach() special-case from sq_quote_buf()
As far as I can tell there has never been a caller of sq_quote_buf()
that expected to invoke it as sq_quote_buf(&buf, buf.buf), which this
code added in 7a33bcbe802 (sq_quote_argv and add_to_string rework with
strbuf's., 2007-09-20) was meant to support. Let's remove that
special-case, none of the other quoting functions behave this way.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>