tests: have "test_hook" be stricter about "mkdir" and "rmdir"
In a preceding commit the "test_hook" wrapper started creating the
.git/hooks directory on-demand, but when that was done a TODO comment
was left behind noting that we should be checking the return value of
"mkdir" and "rmdir".

It's painful to do across the entire test suite because of how some
tests use "--setup", and particularly because "test_when_finished"
queues cleanups in LILO order. Usually that's what's wanted, but when
e.g. doing a hypothetical:

    test_hook foo &&
    test_hook --no-setup-dir bar

It would be convenient if the "rmdir" from the former ran *after* the
"rm" from the former.

Those issues are fixable, but as around 20 tests use "test_hook
--setup" v.s. around 70 that use "test_hook" at all let's implement a
solution that's mostly good enough. We can now provide both
"--no-setup-dir", and change the behavior so that a "--setup" used
anywhere in the test will skip subsequent cleanup of ".git/hooks", as
well making subsequent "mkdir" lax.

Those could be fix, and we could split those two concerns apart, but
by doing this we get more structures for the majority of our tests
that use simple "test_hook" invocations. They'll no longer leave
behind ".git/hook" litter at the end of the test, ensuring that
subsequent tests get a clean slate.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>