Makefile: have "make pot" not "reset --hard"
Change "make pot" to work on copies of the working tree files in a new
".build/pot/" directory, instead of altering them in place and doing a
"reset --hard" afterwards.

This both speeds up the "make pot" target on an initial run, and makes
it *much* faster for incremental runs, and will allow us to implement
related targets in subsequent commits.

When the "pot" target was originally added in cd5513a7168 (i18n:
Makefile: "pot" target to extract messages marked for translation,
2011-02-22) it behaved like a "normal" target. I.e. we'd skip the
re-generation of the po/git.pot if nothing had to be done.

Then after po/git.pot was checked in in dce37b66fb0 (l10n: initial
git.pot for 1.7.10 upcoming release, 2012-02-13) the target was broken
until 1f31963e921 (i18n: treat "make pot" as an explicitly-invoked
target, 2014-08-22) when it was made to depend on "FORCE". I.e. the
Makefile's dependency resolution inherently can't handle incremental
building when the target file may be updated by git (or something else
external to "make").

So we still need a FORCE on po/git.pot, but now we'll have that target
copy ".build/pot/git.pot" to "po/git.pot", and the former doesn't need
FORCE.

I think 1f31963e921 was confused in the rationale for adding "FORCE"
though, it's needed for the reasons noted above, but it's got nothing
to do with how we do C compilation or find headers, which it suggest
is the case. It was also made as a preparatory change for
d85b0dff729 (Makefile: use `find` to determine static header
dependencies, 2014-08-25).

Unlike the compiler, when we generate the "pot "file we won't recurse
from one header file to another, we only consider the specific targets
we're given, so if we miss a bar.h included via foo.h $(LOCALIZED_C)
the FORCE won't help us.

That out of the way, the main logic change here is getting rid of the
"reset --hard":

Before fc0fd5b23bc we'd consider source files as-is with gettext, but
because we need to understand PRItime in the same way that gettext
itself understands PRIuMAX we'd first check if we had a clean
checkout, then munge all of the processed files in-place with "sed",
generate "po/git.pot, and then finally "reset --hard" to undo our
changes.

We didn't need to do this for the *.perl and *.sh files, so let's
skip that and only do it for the *.[ch] files.

Let's then create these munged *.[ch] files in a tree in .build/pot/"
corresponding to our source tree, and have "xgettext" consider
those. The rule needs to be careful to "(cd .build/pot && ...)",
because otherwise the comments in the po/git.pot file wouldn't refer
to the correct source locations (they'd be prefixed with
".build/git/").

We'll then generate intermediate .build/pot/po/% files from %, which
is handy to see at a glance what strings (if any) in a given file are
marked for translation:

	$ make .build/pot/po/pretty.c
	[...]
	$ cat .build/pot/po/pretty.c
	#: pretty.c:1051
	msgid "unable to parse --pretty format"
	msgstr ""
	$

This changes the output of the generated po/git.pot file in one minor
way: Because we're using msgcat(1) instead of xgettext(1) to
concatenate the output we'll now disambiguate where "TRANSLATORS"
comments come from, in cases where a message is the same in N files,
and either only one has a "TRANSLATORS" comment, or they're
different. E.g. for the "Your edited hunk[...]" message we'll now
apply this change (comment content elided):

	+#. #-#-#-#-#  add-patch.c  #-#-#-#-#
	 #. TRANSLATORS: do not translate [y/n]
	[...]
	+#. #-#-#-#-#  git-add--interactive.perl  #-#-#-#-#
	 #. TRANSLATORS: do not translate [y/n]
	[...]
	 #: add-patch.c:1253 git-add--interactive.perl:1244
	 msgid ""
	 "Your edited hunk does not apply. Edit again (saying \"no\" discards!) [y/n]? "
	 msgstr ""

There are three such changes, and they all make the context more
understandable, as msgcat(1) is better at handling these edge cases
than xgettext(1)'s previously used "--join-existing" flag.

The removal of the "--msgid-bugs-address" here and the addition of
"--omit-header" and ".build/pot/pot.header" is because we'll now
always carry forward the existing po/git.pot header.

We'd previously generate the header with xgettext(1), but now we'll be
carrying forward whatever header we find in po/git.pot. Right now this
doesn't matter, we'll even update the timestamp in the header as
before when running "make pot".

But as we'll see in a subsequent commit our previous header behavior
has been hiding an encoding-related bug from us, so let's carry it
forward instead of re-generating it with xgettext(1).

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>