po/git.pot: don't check in result of "make pot"
Remove the "po/git.pot" file from being tracked, which started with
dce37b66fb0 (l10n: initial git.pot for 1.7.10 upcoming release,
2012-02-13).

The reason the po/git.pot started being checked in was because the
po/*.po files were changed a schema where we'd generate them from a
known-good snapshot of po/git.pot, instead of each translator running
"make pot" themselves.

This makes sense, but we don't need to carry this file in-tree just to
achieve that aim, and doing so has resulted in a significant amount of
"diff churn" since this method of doing it was introduced:

    $ git log -p --oneline -- po/git.pot|wc -l
    553743

We can instead have the i18n tooling either read an OID from which to
generate the po/git.pot, or the i18n coordinator could add and commit
a po/git.pot to a branch in his fork.

This reverts to the schema introduced initially in cd5513a7168 (i18n:
Makefile: "pot" target to extract messages marked for translation,
2011-02-22).

The actual "git rm" of po/git.pot is deferred until the subsequent
commit, to make this change easier to review, and to preempt the
mailing list from blocking it due to it being too large.

We'll probably want to adjust the header, but for now it's moved over
as-is from po/git.pot.

Even though this makes .build/po/git.pot strictly redundant to
po/git.pot let's keep the former, and have the rule for the latter
copy it over, just not depending on FORCE now. We still need a
po/*.pot so that "msginit(1)" in the gettext tooling does what we mean
by default, but for other targets we can create things only in
".build" without littering the rest of the working copy.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>