Makefile: always declare [PM]OFILES variables
Change the [PM]OFILES variables to always be declared, and then
continually use them for dependencies depending on NO_GETTEXT.

This untangles an edge-case introduced in 43488240593 (artifacts-tar:
respect NO_GETTEXT, 2021-07-04) and does as I suggested at the time in
[1]. This is needed for a subsequent "check-po" target, which we'll
want to run even if NO_GETTEXT is set.

This also has the advantage that any target we now run that doesn't
require these won't pay the cost of the "wildcard", which was
unconditionally expanded via ":=" instead of lazily via "=".

1. https://lore.kernel.org/git/878s2m8ns9.fsf@evledraar.gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>