po/*.po: don't check in files with --add-location=full
In the preceding commit we manually removed the location information
from the po/*.po files using --no-location. Now we'll avoid committing
this information in-tree, and thus avoid a lot of diff churn, which to
a first approximation is something like:

    $ git log --oneline -p -I'#' -U0 origin/master -- po |wc -l
    1156591
    $ git log --oneline -p -U0 origin/master -- po |wc -l
    3488600

I.e. we'll cut down on the number of lines in our diffs by around 3x,
and for translators make these files much more easy to maintain.

This approach was suggested to me back in [1], but at the time I
either didn't fully understand it, or didn't want to put in the work
to make it happen.

This does introduce the severe regression that anyone working on these
files will no longer get file/line context when translating, which is
why I insisted on including these in the first place, and why the
gettext tooling adds them by default.

I think I was also under the misimpression that when we'd run the
subsequent "msgmerge" as documented in 5e9637c6297 (i18n: add
infrastructure for translating Git with gettext, 2011-11-18) that
having these would help the resulting merger with a new po/git.pot.

Having skimmed gettext-tools/src/msgmerge.c in gettext.git now I don't
think that's the case. It will implicitly behave differently depending
on --sort-output and --sort-by-file, but we're consistent on that
front (by using glob order), so it shouldn't matter for the merge.

But let's first remove the file locations and assert that things stay
that way, in subsequent commits we'll properly support things for
translators again..

1. https://lore.kernel.org/git/AANLkTimGSEVLOjTLt2yntXaDfyUkj=mY46evfRx517MC@mail.gmail.com/
2. https://github.com/git-l10n/git-po

Suggested-by: Erik Faye-Lund <kusmabite@gmail.com>
Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>