i18n: migrate to line-number-less i18n workflow
Change the "flow" of how translators interact with the l10n repository
at [1] to adjust it for a new workflow of not having a po/git.pot file
in-tree at all, and to not commit line numbers to the po/*.po files
that we do track in tree.

The current workflow was added in a combination of dce37b66fb0 (l10n:
initial git.pot for 1.7.10 upcoming release, 2012-02-13) and
271ce198cd0 (Update l10n guide, 2012-02-29).

As noted in preceding commits I think that it came about due to
technical debt I'd left behind in how the "po/git.pot" file was
created, and a mis-impression that the file:line comments were needed
as anything more than a transitory translation aid.

As the updated po/README.md shows the new workflow is substantially
the same, the difference is that translators no longer need to
initially pull from the l10n coordinator for a new po/git.pot, they
can simply use git.git's canonical source repository.

The l10n coordinator is still expected to announce a release to
translate, which presumably would always be Junio's latest release
tag. I'm not certain if this part of the process is actually
important. I.e. the delta translation-wise between that tag and
"master" is usually pretty small, so perhaps translators can just work
on "master" instead.

This also required pulling in some parts of the git-po-helper[1]
logic, in particular the creation of the "core pot" added in its
9efa9a5 (New option --core for init to create core po/pot file,
2021-04-26). As the Makefile changes here show doing so is rather
trivial now that the po/git.pot is generated in the ".build"directory.

We also need to add a new "po-init" target, which replaces both the
"git-po-helper init" and the the instructions for manually munging the
msginit'd file that I added in 5e9637c6297 (i18n: add infrastructure
for translating Git with gettext, 2011-11-18). Now to e.g. create a
new Faroese translation:

    make po-init PO_FILE=po/fo.po

We can furthermore delete the po/pot.header file. We've been relying
on the git.pot to provide a template for us (which msginit implicitly
picks up), but now we don't need to carry it around anymore. We'll
instead munge the msginit'd output appropriately.

1. https://github.com/git-l10n/git-po-helper/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>