ls-tree: add a $(objecttype:padded) format
As noted in the documentation the default ls-tree format was created
at a time when "commit" would never be emitted as an object type, but
ever since that's been the case the OIDs have been mis-aligned in the
output.

It's too late to change that for the default format, but let's provide
an option to do so with --format, for emitting prettier aligned
output.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>