ls-tree API use: pass around "struct repository *" explicitly
Change "ls-tree" to pass around the "struct repository" explicitly,
rather than implicitly relying on "the_repository". This is in
preparation for adding support for recursively listing submodules to
"ls-tree".

To do this some of the tree API needed to be altered to add repo_*()
variants. I didn't bother with adding a repo_parse_tree(), let's have
the one caller in tree.c that needs that call the underlying
repo_parse_tree_gently() directly instead.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>