ls-tree: add a --recurse-submodules option
This is very much a work in progress as the TODO comment indicates,
but it "passes all tests", even if we turn --recurse-submodules on by
default, which isn't scary at all.

It's very much an open question if this will be useful to anyone. I
think for anything to do with OIDs we'll want to add a new format
specifier to give the path to the submodule, so the consumer can know
where to look up what.

Man, architecture kind of sucks, why didn't we stick this all in one
ODB to begin with?

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>