Makefile: avoid "sed" on C files that don't need it
A follow-up to
https://github.com/git-l10n/git-po-helper/issues/12#issuecomment-1090070372,
this does the same thing without the shell-out. We'll even do the
right thing if you "make pot" and add/remove the PRItime string from a
file.

Due to the mentioned caveat with the IMMEDIATE on the LHS of a rule
evaluation this is a bit painful.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>